Fix function calls and reify data segments
Fixed the bug which prevented function calls from other functions.

The problem was caused by a combination of string interning and constant
pool indexes not being shared between chunks.

The solution was to share the same constant pool for all functions
within a compilation unit (which never ends in the REPL).

A persisting problem is in how references to functions get resolved as
global variables, even when they are locally scoped.